Pré-final update
Atualização de métodos, DTO, Controller e Repository. Próximo passo são as medidas de segurança no front e back end

+++ cadastro-alunos/src/main/java/com/example/cadastroalunos/controller/AlunoController.java
@@ -1,13 +1,18 @@
 package com.example.cadastroalunos.controller;
 
+import com.example.cadastroalunos.dto.DadosCadastrarAluno;
+import com.example.cadastroalunos.dto.DadosAtualizarAluno;
 import com.example.cadastroalunos.model.Aluno;
 import com.example.cadastroalunos.repository.AlunoRepository;
 import jakarta.transaction.Transactional;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+
 import java.util.List;
-import java.util.Optional;
+
 
 @RestController
 @RequestMapping("/alunos")
@@ -16,45 +21,59 @@ public class AlunoController {
     @Autowired
     private AlunoRepository repository;
 
-    //Listar
+    //1. Listar TODOS
     @GetMapping
-    public List<Aluno> listar() {
-        return repository.findAll();
+    public ResponseEntity<List<Aluno>> listar() {
+        List<Aluno> lista = repository.findAll();
+        return ResponseEntity.ok().body(lista);
     }
 
+    //1.2 Busca por ID
     @GetMapping("/{id}")
-    public Optional<Aluno> buscarPorId(@PathVariable Long id) {
-        return repository.findById(id);
+    public ResponseEntity<Aluno> buscarPorId(@PathVariable Long id) {
+        var aluno = repository.findById(id);
+
+        //Caso encontre o aluno, deve retornar 20OK, se não, retorna 404 Not Found.
+        return aluno.map(ResponseEntity::ok)
+                .orElse(ResponseEntity.notFound().build());
     }
 
-    //Cadastrar
+    //2. Cadastrar utilizando o DTO e construtor
     @PostMapping
-    public Aluno cadastrar(@RequestBody Aluno aluno) {
-        return repository.save(aluno);
+    @Transactional
+    public ResponseEntity<Aluno> cadastrar(@RequestBody @Valid DadosCadastrarAluno dadosAluno) {
+        //Utiliza o construtor inteligente da entidade Aluno.
+        Aluno aluno = new Aluno(dadosAluno);
+        repository.save(aluno);
+
+        return ResponseEntity.ok().body(aluno);
     }
 
-    //Atualizar
+    //3. Atualizar utilizando mesma base da função cadastrar
     @PutMapping("/{id}")
     @Transactional
-    public Aluno atualizar(@PathVariable Long id,
-                           @RequestBody Aluno alunoAtualizado) {
+    public ResponseEntity<Aluno> atualizar(@PathVariable Long id,
+                                           @RequestBody @Valid DadosAtualizarAluno dadosAluno) {
+        //Busca o aluno no banco ou lança uma exceção se não encontrar
+        Aluno aluno = repository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Aluno não encontrado"));
 
-        Aluno aluno = repository.findById(id).orElseThrow();
+        //Executa as verificações de dados Null
+        aluno.atualizarAluno(dadosAluno);
 
-        aluno.setNome(alunoAtualizado.getNome());
-        aluno.setMatricula(alunoAtualizado.getMatricula());
-        aluno.setEmail(alunoAtualizado.getEmail());
-        aluno.setCurso(alunoAtualizado.getCurso());
-        aluno.setTelefone(alunoAtualizado.getTelefone());
-        aluno.setEndereco(alunoAtualizado.getEndereco());
-        aluno.setDataMatricula(alunoAtualizado.getDataMatricula());
-
-        return repository.save(aluno);
+        //Por utilizar o Transactional é desnecessário utilizar repository.save(aluno).
+        //O JPA detecta a mudança e atualiza o banco sozinho ao final do méteodo.
+        return ResponseEntity.ok(aluno);
     }
 
-    //Delete
+    //4. Deletar
     @DeleteMapping("/{id}")
-    public void deletar(@PathVariable Long id) {
+    @Transactional
+    public ResponseEntity<Void> deletar(@PathVariable Long id) {
+        if (!repository.existsById(id))  {
+            return ResponseEntity.notFound().build();
+        }
         repository.deleteById(id);
+        return ResponseEntity.noContent().build(); //Deve retornar Status 204 No Content
     }
 }
